backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0